refactor(web): step 11a/11c — extract useComposerAttachments and useComposerDrafts from chat-pane
